+++ src/test/java/io/beanmapper/testmodel/optional_getter/MyEntityResult.java
@@ -1,0 +1,9 @@
+package io.beanmapper.testmodel.optional_getter;
+
+public class MyEntityResult {
+
+    public String value;
+
+    public MyEntityResult child;
+
+}